requirements.txt
annotated-types==0.7.0
anyio==3.7.1
APScheduler==3.11.0
asgiref==3.8.1
Authlib==1.4.1
azure-common==1.1.28
azure-core==1.32.0
azure-identity==1.20.0
azure-keyvault-keys==4.10.0
azure-mgmt-core==1.5.0
azure-mgmt-hardwaresecuritymodules==1.0.0b2
babel==2.17.0
bcrypt==4.2.1
blinker==1.9.0
boto3==1.36.24
botocore==1.36.24
buildozer==1.5.0
CacheControl==0.14.2
cachetools==5.5.1
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.8
cloudinary==1.44.1
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.1
cycler==0.12.1
Cython==3.0.11
distlib==0.3.8
Django==5.1.6
dnspython==2.7.0
docutils==0.21.2
ecdsa==0.19.1
email_validator==2.2.0
et_xmlfile==2.0.0
fastapi==0.103.1
fastapi-limiter==0.1.6
filelock==3.16.1
firebase-admin==6.2.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-DebugToolbar==0.16.0
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
fonttools==4.55.3
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.1
google-cloud-firestore==2.20.0
google-cloud-kms==3.3.1
google-cloud-storage==3.0.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
greenlet==3.2.2
grpc-google-iam-v1==0.14.0
grpcio==1.70.0
grpcio-status==1.70.0
gunicorn==21.2.0
h11==0.16.0
httplib2==0.22.0
idna==3.10
iniconfig==2.0.0
isodate==0.7.2
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.6
jmespath==1.0.1
keyring==25.6.0
Kivy==2.3.0
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.7.0
Kivy-Garden==0.1.5
kivy_garden.matplotlib==0.1.1.dev0
kiwisolver==1.4.8
liboqs-python @ git+https://github.com/open-quantum-safe/liboqs-python.git@1a29e9342182c829f4fc2ff47b5366796c7ffb64
MarkupSafe==3.0.2
matplotlib==3.10.0
more-itertools==10.6.0
MouseInfo==0.1.3
msal==1.31.1
msal-extensions==1.2.0
msgpack==1.1.0
mss==10.0.0
mysql-connector-python==9.4.0
numpy==2.2.1
oauthlib==3.2.2
opencv-python==4.10.0.84
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
oqs==0.10.2
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
plotly==5.24.1
pluggy==1.5.0
plyer==2.1.0
portalocker==2.10.1
proto-plus==1.26.0
protobuf==5.29.3
psutil==6.1.1
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycparser==2.22
pycryptodome==3.21.0
pydantic==1.10.22
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydub==0.25.1
pygame==2.6.1
PyGetWindow==0.0.9
Pygments==2.18.0
pyjnius==1.6.1
PyJWT==2.10.1
PyKCS11==1.5.17
PyMsgBox==1.0.9
pynput==1.7.7
pyotp==2.9.0
pyparsing==3.2.1
pyperclip==1.9.0
pypiwin32==223
PyRect==0.2.0
PyScreeze==1.0.1
pytesseract==0.3.13
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.5.0
python-multipart==0.0.20
pytweening==1.2.0
pytz==2024.2
pywin32==306
pywin32-ctypes==0.2.3
redis==6.0.0
reportlab==4.3.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
s3transfer==0.11.2
schedule==1.2.2
seaborn==0.13.2
setuptools==75.8.0
sh==2.0.7
six==1.17.0
sniffio==1.3.1
SpeechRecognition==3.14.1
SQLAlchemy==2.0.40
sqlparse==0.5.3
starlette==0.27.0
tenacity==9.0.0
tinydb==4.8.2
tkcalendar==1.6.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2024.2
tzlocal==5.3
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.35.0
virtualenv==20.26.6
Werkzeug==3.1.3
wheel==0.45.1
